refactor action functions
