+++ go.mod
@@ -2,8 +2,9 @@ module github.com/Ygg-Drasill/Sleipnir
 
 go 1.22
 
+require github.com/spf13/cobra v1.8.0
+
 require (
 	github.com/inconshreveable/mousetrap v1.1.0 // indirect
-	github.com/spf13/cobra v1.8.0 // indirect
 	github.com/spf13/pflag v1.0.5 // indirect
 )
